backend/requirements.txt
django>=5.0,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-allauth>=0.61
anthropic>=0.25
Pillow>=10.0
psycopg2-binary>=2.9
python-dotenv>=1.0
requests>=2.31
gunicorn>=21.0
dj-database-url>=2.1
whitenoise>=6.7
celery[redis]>=5.3
redis>=5.0
